Add files via upload
Hey guys! I uploaded the preliminary schema. I used the @name attribute from the <character> element as the identifier of <speaker> elements as well, so we'll have an easy point of reference between the two. For example, if you wanted to search for <speaker> elements after mark up, it would give you the @name attribute and the actual text of the name. The @name attribute could then be cross-referenced with the <character> element by searching the @name to examine the rest of the data, such as @gender and @class. 

Also, I write my schemas linearly and use the full words for element names (ie, <paragraph> instead of <p>), but if we want to change this later for mark up we can. Test it out and let me know if there's anything I got wrong or you think we should change!