Add files via upload
FYI I made <scary_word> a REQUIREMENT within the <scary_phrase> element, that way we wouldn't forget to tag the within the phrases. This assumes that a phrase must have at least one scary word in it to make it a scary phrase, which I think is a good general rule of thumb. Let me know if you disagree/run into trouble.