John C. markup and schema fixes
1.  I made a (possibly wrong) decision to use speaker elements as
milestones.  In my markup, they precede every speech element.  This was
pretty easy using regex.  Once you do that, the schema requires you to
add a name attribute to each speaker tag, which is great because you can
find and fix them (manually i am afraid).
2. I used "amb" as a speaker tag attribute a few times, when a mixed
group spoke or speaker was not specified.
3. I haven't done the scary markup yet, this is just structural stuff
4. I saved this as a different version than the original, mostly raw
text just in case we need to overhaul.
5. Make sure you tag all of your speech before you do anything else, as
attributes will make it impossible to automate later on (I learned this
the hard way)